Add missing attributes for theming

+++ thememanager/src/main/java/company/tap/thememanager/theme/BaseTextTheme.kt
@@ -6,6 +6,6 @@ package company.tap.thememanager.theme
  */
 open class BaseTextTheme {
     var textColor: Int? = null
-    var textSize: Float? = null
-    var letterSpacing: Float? = null
+    var textSize: Double? = null
+    var letterSpacing: Double? = null
 }
